Refactor de les noves Entitats v.2
-Noves entitats creades, pendent de debugar
-Converses, productes comprats, productes venuts d'un usuari pendents

+++ app/src/main/java/org/udg/pds/cheapyandroid/entity/Categoria.java
@@ -14,7 +14,7 @@ import java.util.Map;
 public class Categoria {
 
     @JsonProperty("id")
-    private long id;
+    private Long id;
     @JsonProperty("nom")
     private String nom;
     @JsonIgnore
@@ -26,7 +26,7 @@ public class Categoria {
     }
 
     @JsonProperty("id")
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
